Top level users fetch, user profile fetch

+++ app/src/main/java/com/l0mtick/founditmobile/main/presentation/home/components/TopLevelUsersRow.kt
@@ -33,7 +33,7 @@ import com.l0mtick.founditmobile.ui.theme.Theme
 fun TopLevelUsersRow(
     users: List<User>,
     modifier: Modifier = Modifier,
-    onUserCardClick: (Long) -> Unit
+    onUserCardClick: (Int) -> Unit
 ) {
     Row(
         modifier = modifier
